requirements.txt
streamlit
requests
beautifulsoup4
python-dotenv
google-generativeai
openai
# Optional (for parsing domain names or advanced analysis)
# tldextract